Attack works(correctly)
The problem was that in the update method Update, we were using the transform.position.x and transform.position.y instead of transform.parent.position, this meant that the x and y coordinates of the swing were adding to themselves instead of to the parent coordinates, this made the coordinates of the swing get off the screen. To explain the previous, if the player x coordinates is 3, and we have a localX of 1, we want the coordinates of the swing to be either 2 or 4, depending to were we look, so thats why we use transform.parent, to get that 3, because if we use transform without the .parent we will be getting the coodinate of the swing, which if we always look to the right would make the swing w coordinate go(for example)->4,5,6,7,8,9,10,... . This because we keep operating up on itself, thats why we need to add or subtract relative to the parent position. Idk if i explained it perfectly, i hope its understandable