Controls disabled when dead
When playing again, even tho it worked perfectly, it displayed errors because the program was trying to call to the functions of the players we deleted (for example, if the original VirtualGuy its subscribed to jump, if we play again the new player will also be subscribed to jump, so when jump happens it will call both the new player and the old virtual guy(who is actually destroyed), and that gave a warning). Its not that deep of a problem, just remember that when using the input system, the same way you subscribe the players to certain actions, you need to unsuscribe then (disable their controlls) when they die.