Hovering over a player selector now shows info
In the character selection screen, when hovering the mouse over a button, this would make that a text associated to the button will be visible while we hover over it. To change the text go into the TextWindow GameObjects inside of the buttons. In the future we could change the script to have access to player stats instead of only a text if we want to. IMPORTANT: when changing the text do it in the copies inside the grid in character selector, not in the prefab they all share, because in that case they will all have the same text