Added Cooldown bars
Still not complete, but the basics are there and easy to understand. Still need to add a cooldown display for the dash (its easy to do, but i wanna take a break), and also decorate the progress bars, to actually look good ingame, because right now they maybe are too simple. As of rigth now, the yellow bar represents the basic attack, and the green bar represents the special attack