Lick and Charge attacks added + fixed spinning attack hitbox
Neither do Lick, Charge or SpinnnigAttack have animations, but the attacks work. Lick goes the way the player looks, and slows the enemy for 4s apart from dealing damage. Charge is basically a faster dash that does big damage, a good contrast for a character thats actually slow, reminds me of a american football player charge if u want to use that analogy. The special thing about Spinning attack is that its the only attack that goes right and left at the same time. Lastly, i need to implement the throwing spear attack for the chaman, that as the name says, what will make it special its that itll be a throwable attack.